Adds logic to create DB records to begin building a local cache of price history

+++ source/api/db/migrate/20140211201203_create_trades.rb
@@ -3,9 +3,9 @@ class CreateTrades < ActiveRecord::Migration
     create_table :trades do |t|
       t.belongs_to :market
       t.integer :transaction_id
-      t.decimal :btc_price_per_coin
-      t.decimal :quantity
-      t.decimal :btc_total
+      t.float :btc_price_per_coin
+      t.float :quantity
+      t.float :btc_total
       t.datetime :executed_at
 
       t.timestamps
